fix(daemon): share distill validation and read helpers
